refactor(ui): extract view-model/internal.ts + view-model/device-names.ts
Move the private text/peer-error helpers (cleanText, normalizeDisplayName,
friendlyDeviceFallback, peerErrorText, isUnauthorizedPeerError,
isConnectivityPeerError) into view-model/internal.ts, and the
device-name derivations (resolveFriendlyDeviceName, deviceNeedsFriendlyName)
into view-model/device-names.ts. Also drop the unused _looksLikeDeviceId
dead helper. The barrel re-exports the two device-name functions so
existing call sites stay on the same import path.
